requirements.txt
# Core dependencies
faster-whisper>=0.9.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.20.0
sounddevice>=0.4.5
PyQt6>=6.4.0
pyaudio>=0.2.13
python-dotenv>=1.0.0
pydub>=0.25.1
xdotool>=0.0.1
pyperclip>=1.8.2
python-xlib>=0.33
pynput>=1.7.6
transformers>=4.28.0
ffmpeg-python>=0.2.0
